+++ src/main/java/logic/CreateurPlanete.java
@@ -3,7 +3,7 @@ package logic;
 public class CreateurPlanete extends CreateurAstre{
 
     @Override
-    public Planete factory(String nom, float taille, float masse, int pX, int pY){
+    public Planete factory(String nom, int taille, int masse, int pX, int pY){
         return new Planete(nom, taille, masse, pX, pY);
     }
 }
